Moved base manager to /connections
