+++ aries_cloudagent/protocols/didexchange/v1_0/tests/test_manager.py
@@ -22,7 +22,7 @@ from ....out_of_band.v1_0.models.invitation import InvitationRecord
 from .....multitenant.manager import MultitenantManager
 from .....wallet.base import DIDInfo
 from .....wallet.in_memory import InMemoryWallet
-from ....connections.v1_0.base_manager import (
+from .....connections.base_manager import (
     BaseConnectionManager,
     BaseConnectionManagerError,
 )
